requirements.txt
# Core Dependencies
python-dotenv>=1.0.0

# Image Processing
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# OCR Engines
easyocr>=1.7.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# API Clients
groq>=0.4.0

# Document Generation
python-docx>=1.0.0

# Web Interface
streamlit>=1.28.0

# Optional: For GPU support (uncomment if using CUDA)
# torch==2.0.0+cu118 --extra-index-url https://download.pytorch.org/whl/cu118
